[rtp sender] switch back to an executor (fixes: #80)
While investigating #80 it turns out the problem was not the time spend
performing VP8 encoding, but the time between the end of encoding and
the main thread receiving the results.

Using a worker thread and a Future gave completely random times, up to
tens of milliseconds. This resulted in the encoding loop falling behind
massively, even with a dummy VideoStreamTrack. Reverting to asyncio's
run_in_executor seems to consistently give sub-millisecond times.