If all DTLS transports are closed, close the peer connection
When the remote peer closes the peer connection, API users need to wait
until the ICE consent expires for the peer connection to move to the
"closed" state.

However, we will usually receive a DTLS close_notify immediately from
the remote peer. If all DTLS transports are closed, shutdown the peer
connection.

Fixes: #934